+++ CHANGELOG.md
@@ -24,6 +24,10 @@ Quality:
 
 - Overlapping characters in text now render correctly. (@Xertes0)
 
+Usability:
+
+- touchHLE now supports real accelerometer input on devices with a built-in accelerometer, such as phones and tablets. This is only used if no game controller is connected. (@hikari-no-yume)
+
 Other:
 
 - touchHLE now has a primitive implementation of the GDB Remote Serial Protocol. GDB can connect to touchHLE over TCP and set software breakpoints, inspect memory and registers, step or continue execution, etc. This replaces the old `--breakpoint=` option, which is now removed. (@hikari-no-yume)